Add instructions how to get Xcode project that can build
docs: Add instructions how to get Xcode project that can build

If current project is opened on a different machine, it will not be
able to build due to use of absolute paths in the cmake Xcode
generator. Thus suggesting to re-generate the Xcode project in order to
get project that can build.